pyproject.toml
[project]
name = "modeling-framework"
version = "0.1.0"
description = "A modular and reproducible framework for simulating, fitting and analyzing compartmental epidemic models"
authors = [{ name = "Shreya Mukherjee" }]
requires-python = ">=3.11"

dependencies = [
    "numpy",
    "pandas",
    "geopandas",
    "matplotlib",
    "scikit-learn",
    "typer",
    "tqdm",
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pre-commit",
]
docs = [
    "pymdown-extensions",
    "mkdocstrings-python",
    "mkdocs-material",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/epimodels"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "FLY", "RUF", "B", "ICN"]
ignore = []

[tool.ruff]
line-length = 140
indent-width = 4

[tool.ruff.isort]
known-first-party = ["epimodels"]
